Add support for VMWare
Working with VMWare virtual machines is now supported. To create the
machines using VMWare, use one of the following commands:

  vagrant up --provider vmware_fusion
  vagrant up --provider vmware_workstation

The VirtualBox provider will still work, and can be used by not
specifying a provider.

The base boxes were also switched to a slightly older version of Ubuntu
so that there would be a matching version between providers. The
provisioning was also changed slightly to account for this