Provision using ansible instead of chef.
Chef always seemed too complex and I much prefer ansible playbooks for
provisioning. While the playbooks don't quite do as much as the old
cookbooks yet, they seem much faster.

Additionally I should be able to iterate on the roles a little bit and
create a playbook for provisioning a linux or OSX machine on the host
rather than in a guest. A lot of the Review Board environment specific
stuff could be reused for this.

I also decided to take care of cloning the Review Board code and
completely setting up the development environment as part of
provisioning. If a more custom environment is needed it can be done
manually after provisioning, but this is a much easier starting point
for one off development or students.